ai-nl-analytics-dashboard/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
pandas==2.2.3
numpy==2.1.3
chardet==5.2.0
sqlparse==0.5.1
google-generativeai==0.8.3

